updated many2many relation
