+++ src/Entity/Author.php
@@ -25,11 +25,6 @@ class Author
     private $name;
 
     /**
-     * @ORM\OneToMany(targetEntity=Book::class, mappedBy="author")
-     */
-    private $books;
-
-    /**
      * @ORM\Column(type="integer")
      */
     private $age;
@@ -39,9 +34,11 @@ class Author
      */
     private $country;
 
+
+
     public function __construct()
     {
-        $this->books = new ArrayCollection();
+
     }
 
 
@@ -58,36 +55,6 @@ class Author
     public function setName(string $name): self
     {
         $this->name = $name;
-
-        return $this;
-    }
-
-    /**
-     * @return Collection|Book[]
-     */
-    public function getBooks(): Collection
-    {
-        return $this->books;
-    }
-
-    public function addBook(Book $book): self
-    {
-        if (!$this->books->contains($book)) {
-            $this->books[] = $book;
-            $book->setAuthor($this);
-        }
-
-        return $this;
-    }
-
-    public function removeBook(Book $book): self
-    {
-        if ($this->books->removeElement($book)) {
-            // set the owning side to null (unless already changed)
-            if ($book->getAuthor() === $this) {
-                $book->setAuthor(null);
-            }
-        }
 
         return $this;
     }
