+++ templates/admin/book.html.twig
@@ -65,14 +65,23 @@
       </div>
 
   <div class="tab-content">
-   <div class="tab-content__pane active" id="classes">
-    {% for author, book_records in books %}
-     <div class="mt-3 mb-2">
-      <h4 class="text-xl text-gray-700 dark:text-gray-600 font-medium leading-none mt-3">
-       {{ author|upper }}
-      </h4>
-     </div>
-     {% for  key,book in book_records %}
+
+     {% for book in books %}
+
+      Authors :
+   {% for tag in book.bookTags %}
+
+    ({{ tag.author }})
+   {% endfor %}
+
+
+      Tags :
+      {% for tag in book.bookTags %}
+
+       ({{ tag.tag }})
+      {% endfor %}
+
+
       {% set color = 'text-theme-9' %}
       <div class="intro-y mt-1">
        <div class="box grid grid-cols-8  gap-1 mt-5 toggle_details zoom-in">
@@ -113,10 +122,8 @@
        </div>
       </div>
      {% endfor %}
-    {% endfor %}
     {#                {{ include('/admin/partials/pagination.html.twig') }}#}
-   </div>
-  </div>
+
  </div>
 {% endblock %}
 {% block javascript %}
